the questionair

+++ src/main/java/com/insurance/aml/entity/AmlQuestion.java
@@ -10,6 +10,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -66,7 +67,7 @@ public class AmlQuestion {
     private LocalDateTime updatedAt;
 
     @Column(name = "version", nullable = false)
-    private Float version;
+    private BigDecimal version;
 
     @Builder.Default
     @OneToMany(mappedBy = "question", cascade = CascadeType.ALL, orphanRemoval = true, fetch = FetchType.LAZY)
